requirements.txt
certifi==2020.6.20
chardet==3.0.4
urllib3==1.24.3
idna==2.8
requests==2.21.0
schedule==0.6.0
selenium==3.141.0
PyAutoGUI==0.9.52
webdriver-manager==3.2.2
Discord-Webhooks==1.0.4
opencv-python==4.4.0.46
imutils==0.5.4
Pillow==8.0.1
python-dotenv==0.15.0